{DBA-1} Change project structure
