+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "angular-todo",
+  "name": "angular-web",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
