--- src/app/app.component.html
@@ -1,16 +1,0 @@
-@if (true) {
-    <app-header></app-header>
-    <router-outlet/>
-} @else {
-    @defer {
-        <div class="error-wrapper">
-            <span class="error">Все сломано :(</span>
-        </div>
-    } @loading (minimum 2000) {
-        <div class="loader-wrapper">
-            <span class="loader">Загрузка...</span>
-        </div>
-    }
-}
-
-<!--  Просто экспериментировал с defer. СМотрел как в DBA это работало    -->
