--- src/app/app.component.ts
@@ -1,13 +1,0 @@
-import {Component} from '@angular/core';
-import {RouterOutlet} from '@angular/router';
-import {HeaderComponent} from "./shared/components/header/header.component";
-
-@Component({
-    selector: 'app-root',
-    imports: [RouterOutlet, HeaderComponent],
-    templateUrl: './app.component.html',
-    styleUrl: './app.component.scss'
-})
-export class AppComponent {
-    title = 'angular-posts';
-}
