--- src/app/base/home/home.component.ts
@@ -1,12 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-    selector: 'app-home',
-    standalone: true,
-    imports: [],
-    templateUrl: './home.component.html',
-    styleUrl: './home.component.scss'
-})
-export class HomeComponent {
-
-}
