--- src/app/core/api/posts.api.service.ts
@@ -1,10 +1,0 @@
-import {Injectable} from '@angular/core';
-
-@Injectable({
-    providedIn: 'root'
-})
-export class PostsApiService {
-
-    constructor() {
-    }
-}
