--- src/app/shared/components/header/header.component.ts
@@ -1,16 +1,0 @@
-import {Component} from '@angular/core';
-import {RouterLink, RouterLinkActive} from "@angular/router";
-import {ROUTE_PATHS} from "../../constants/route-paths";
-
-@Component({
-    selector: 'app-header',
-    imports: [
-        RouterLink,
-        RouterLinkActive
-    ],
-    templateUrl: './header.component.html',
-    styleUrl: './header.component.scss'
-})
-export class HeaderComponent {
-    ROUTE = ROUTE_PATHS;
-}
