+++ src/apps/en/src/app/base/app.component.ts
@@ -1,0 +1,13 @@
+import {Component} from '@angular/core';
+import {RouterOutlet} from '@angular/router';
+import {HeaderComponent} from "../ui/dumb/header/header.component";
+
+@Component({
+    selector: 'apps-root',
+    imports: [RouterOutlet, HeaderComponent],
+    templateUrl: './app.component.html',
+    styleUrl: './app.component.scss'
+})
+export class AppComponent {
+    title = 'angular-posts';
+}
