+++ src/apps/en/src/app/base/home/home.component.ts
@@ -1,0 +1,12 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'apps-home',
+    standalone: true,
+    imports: [],
+    templateUrl: './home.component.html',
+    styleUrl: './home.component.scss'
+})
+export class HomeComponent {
+
+}
