+++ src/apps/en/src/app/core/common/styles/mixins.scss
