+++ src/apps/en/src/app/core/common/styles/styles.scss
