+++ src/apps/en/src/app/infrastructure/configs/app.config.ts
@@ -1,0 +1,13 @@
+import {ApplicationConfig, provideZoneChangeDetection} from '@angular/core';
+import {provideRouter} from '@angular/router';
+import {routes} from '../../core/routes';
+import {provideHttpClient} from "@angular/common/http";
+
+
+export const appConfig: ApplicationConfig = {
+    providers: [
+        provideZoneChangeDetection({eventCoalescing: true}),
+        provideRouter(routes),
+        provideHttpClient()
+    ]
+};
