+++ src/apps/en/src/app/ui/dumb/header/header.component.ts
@@ -1,0 +1,16 @@
+import {Component} from '@angular/core';
+import {RouterLink, RouterLinkActive} from "@angular/router";
+import {ROUTE_PATHS} from "../../../core/common/constants/route-paths";
+
+@Component({
+    selector: 'apps-header',
+    imports: [
+        RouterLink,
+        RouterLinkActive
+    ],
+    templateUrl: './header.component.html',
+    styleUrl: './header.component.scss'
+})
+export class HeaderComponent {
+    ROUTE = ROUTE_PATHS;
+}
