+++ tsconfig.app.json
@@ -7,7 +7,7 @@
     "types": []
   },
   "files": [
-    "src/main.ts"
+      "src/apps/en/src/main.ts"
   ],
   "include": [
     "src/**/*.d.ts"
